+++ src/main/java/com/tenio/examples/example1/TestClientLogin.java
@@ -24,7 +24,9 @@ THE SOFTWARE.
 
 package com.tenio.examples.example1;
 
-import com.tenio.common.data.utility.ZeroUtility;
+import com.tenio.common.data.DataType;
+import com.tenio.common.data.DataUtility;
+import com.tenio.common.data.zero.utility.ZeroUtility;
 import com.tenio.core.entity.data.ServerMessage;
 import com.tenio.examples.client.ClientUtility;
 import com.tenio.examples.client.SocketListener;
@@ -69,7 +71,10 @@ public final class TestClientLogin implements SocketListener {
   }
 
   @Override
-  public void onReceivedTCP(ServerMessage message) {
+  public void onReceivedTCP(byte[] binaries) {
+    var dat = DataUtility.binaryToCollection(DataType.ZERO, binaries);
+    var message = ServerMessage.newInstance().setData(dat);
+
     System.out.println("[RECV FROM SERVER TCP] -> " + message.getData().toString());
 
     try {