+++ src/main/java/com/tenio/examples/example10/TestServerMsgPackEcho.java
@@ -22,27 +22,18 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 THE SOFTWARE.
 */
 
-package com.tenio.examples.example3.handler;
+package com.tenio.examples.example10;
 
-import com.tenio.common.bootstrap.annotation.Component;
-import com.tenio.common.data.zero.ZeroMap;
-import com.tenio.core.entity.Player;
-import com.tenio.core.entity.data.ServerMessage;
-import com.tenio.core.handler.AbstractHandler;
-import com.tenio.core.handler.event.EventReceivedMessageFromPlayer;
-import com.tenio.examples.server.SharedEventKey;
+import com.tenio.common.bootstrap.annotation.Bootstrap;
+import com.tenio.core.ApplicationLauncher;
 
-@Component
-public final class ReceivedMessageFromPlayerHandler extends AbstractHandler
-    implements EventReceivedMessageFromPlayer {
+/**
+ * This class shows how a server handle messages that came from a client.
+ */
+@Bootstrap
+public final class TestServerMsgPackEcho {
 
-  @Override
-  public void handle(Player player, ServerMessage message) {
-    var data =
-        map().putString(SharedEventKey.KEY_CLIENT_SERVER_ECHO, String.format("Echo(%s): %s",
-            player.getName(),
-            ((ZeroMap) message.getData()).getString(SharedEventKey.KEY_CLIENT_SERVER_ECHO)));
-
-    response().setContent(data.toBinary()).setRecipientPlayer(player).prioritizedUdp().write();
+  public static void main(String[] params) {
+    ApplicationLauncher.run(TestServerMsgPackEcho.class, params);
   }
 }