+++ src/main/java/com/tenio/examples/example10/handler/AttachConnectionRequestValidatedHandler.java
@@ -22,27 +22,25 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 THE SOFTWARE.
 */
 
-package com.tenio.examples.example3.handler;
+package com.tenio.examples.example10.handler;
 
 import com.tenio.common.bootstrap.annotation.Component;
-import com.tenio.common.data.zero.ZeroMap;
+import com.tenio.common.data.msgpack.element.MsgPackMap;
 import com.tenio.core.entity.Player;
 import com.tenio.core.entity.data.ServerMessage;
 import com.tenio.core.handler.AbstractHandler;
-import com.tenio.core.handler.event.EventReceivedMessageFromPlayer;
+import com.tenio.core.handler.event.EventAttachConnectionRequestValidation;
 import com.tenio.examples.server.SharedEventKey;
+import java.util.Optional;
 
 @Component
-public final class ReceivedMessageFromPlayerHandler extends AbstractHandler
-    implements EventReceivedMessageFromPlayer {
+public final class AttachConnectionRequestValidatedHandler extends AbstractHandler
+    implements EventAttachConnectionRequestValidation {
 
   @Override
-  public void handle(Player player, ServerMessage message) {
-    var data =
-        map().putString(SharedEventKey.KEY_CLIENT_SERVER_ECHO, String.format("Echo(%s): %s",
-            player.getName(),
-            ((ZeroMap) message.getData()).getString(SharedEventKey.KEY_CLIENT_SERVER_ECHO)));
+  public Optional<Player> handle(ServerMessage message) {
+    var data = (MsgPackMap) message.getData();
 
-    response().setContent(data.toBinary()).setRecipientPlayer(player).prioritizedUdp().write();
+    return api().getPlayerByName(data.getString(SharedEventKey.KEY_PLAYER_LOGIN));
   }
 }