+++ src/main/java/com/tenio/examples/example6/handler/PlayerLoggedInHandler.java
@@ -38,7 +38,7 @@ public final class PlayerLoggedInHandler extends AbstractHandler
   @Override
   public void handle(Player player, PlayerLoggedInResult result) {
     if (result == PlayerLoggedInResult.SUCCESS) {
-      var data = object()
+      var data = map()
           .putString(SharedEventKey.KEY_CLIENT_SERVER_ECHO,
               String.format("Welcome to server: %s", player.getName()))
           .putString(SharedEventKey.KEY_PLAYER_LOGIN, player.getName());