+++ src/main/java/com/tenio/examples/example7/handler/CcuInformationHandler.java
@@ -33,6 +33,7 @@ public final class CcuInformationHandler extends AbstractHandler implements Even
 
   @Override
   public void handle(int numberPlayers) {
-    info("CCU INFO", String.format("Current players: %d", numberPlayers));
+    info("CCU INFO", String.format("Current players: %d, Uptime: %d seconds", numberPlayers,
+        api().getUptime() / 1000L));
   }
 }