+++ src/main/java/com/tenio/examples/example9/handler/AttachConnectionRequestValidatedHandler.java
@@ -25,7 +25,7 @@ THE SOFTWARE.
 package com.tenio.examples.example9.handler;
 
 import com.tenio.common.bootstrap.annotation.Component;
-import com.tenio.common.data.ZeroMap;
+import com.tenio.common.data.zero.ZeroMap;
 import com.tenio.core.entity.Player;
 import com.tenio.core.entity.data.ServerMessage;
 import com.tenio.core.handler.AbstractHandler;