+++ src/main/java/com/tenio/examples/example9/handler/AttachedConnectionHandler.java
@@ -40,7 +40,7 @@ public final class AttachedConnectionHandler extends AbstractHandler
   @Override
   public void handle(Optional<Player> player, int kcpConv, AttachedConnectionResult result) {
     if (result == AttachedConnectionResult.SUCCESS) {
-      var data = object().putZeroArray(SharedEventKey.KEY_ALLOW_TO_ATTACH,
+      var data = map().putZeroArray(SharedEventKey.KEY_ALLOW_TO_ATTACH,
           array().addByte(UdpEstablishedState.ATTACHED).addInteger(kcpConv));
 
       response().setContent(data.toBinary()).setRecipientPlayer(player.get()).write();