+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## 0.0.1 2021-10-15 Feature release
+### 0.0.1 2021-10-15 Feature release
 **Initial release**
 - Login example  
 - FSM example  
@@ -9,7 +9,7 @@
 - Websocket example  
 - HTTP example  
 
-## 0.2.0 2022-04-20 Feature release
+### 0.2.0 2022-04-20 Feature release
 - Updated the framework dependencies  
 
 ### 0.3.0 2022-10-14 Feature release
@@ -18,4 +18,11 @@
 - Refactoring: renamed package "extension" to "handler"
 - Allowed declaring the server address in configuration file
 - Supported KCP transportation
+- Bugs fixed
+
+### 0.3.1 2022-10-21 Feature release
+- Added new example to show how to use MsgPack
+- Added new events to handle connection-refused cases
+- WebSocket handling issues fixed
+- Allowed showing server's uptime
 - Bugs fixed